Extract the E-L* synthesis loop out of lstar.py
lstar.py welded two unrelated things together: the shared classification and
accuracy primitives (_oracle_classify, denoise_accept_labels, locate_incorrect_point,
_batch_before_possible_stop, estimate_agreement_rate), and the E-L* counterexample-
driven synthesis algorithm. Move the latter into counterexample_synthesis.py, leaving
lstar.py as just the substrate.

Pure move -- function bodies are byte-for-byte, only the module docstrings and import
lists change. learn.py and test_lstar import the loop from its new home.
